Fetch API's for Replenishment for all hierarchy

+++ src/main/java/com/walmart/aex/sp/dto/buyquantity/MetricsDto.java
@@ -14,7 +14,7 @@ public class MetricsDto {
     private Integer finalReplenishmentQty;
     private Integer vendorPack;
     private Integer warehousePack;
-    private Integer packRatio;
+    private Double packRatio;
     private Integer replenishmentPacks;
 
 }
